8 Queens Logic works
Figured out an algorithm to correct display safe boxes when placing
queens. It is definitely a "brute force" way and is not very clever, but
the main task for this assignment was not the algorithm but rather using
C# to draw the board, update hints and queens correctly, incorporate
sounds and different modes, etc.